code/requirements.txt
Flask==1.1.2
markdownify==0.5.1
marshmallow==3.7.1
requests==2.24.0
zappa==0.51.0
pyjwt[crypto]==2.0.0
flake8==3.8.3
coverage==5.2.1
pytest==6.0.1
